overhauled input handlers & header
Put event listeners on main & header only (using revamped input fn to handle all types of events).

Also, made the left-most button in the header change to an options button if no  items are selected and showed the selection path in the search field if items are selected instead of just "search".