partial bug fix
If some items are selected, input.repeat now only uncompltes the selected items, & not all of their sibilings.

This is only a partial fix as more work will need to be done on logic.repeat, but desired behavior must be decided upon ahead of time (99% sure how it should work, but don't want to rush into it at the moment in case I'm overlooking something).